model auto-construction
The simulator, loads the styles of aerodynamics and propulsion form parameter server
