+++ last_letter_2/include/dynamics.hpp
@@ -1,0 +1,10 @@
+class Dynamics
+{
+  public:
+  Model * model;
+  Aerodynamics *aerodynamics;
+  Propulsion * propulsion;
+  Dynamics(Model *);
+  void calcAero();
+  void calcProp();
+};
