+++ last_letter_2/include/factory.hpp
@@ -1,0 +1,6 @@
+class Factory
+{
+    public:
+    Aerodynamics * buildAerodynamics(Model *);
+    Propulsion * buildPropulsion(Model *);
+};
