+++ last_letter_2/include/last_letter_2_libs.hpp
@@ -21,3 +21,4 @@ class Propulsion;
 class Dynamics;
 class Model;
 class Master;
+class Factory;
