+++ last_letter_2/src/propulsion.cpp
@@ -2,9 +2,7 @@
 Propulsion::Propulsion(Model *parent)
 {
     model=parent;
-    initParam();
 }
-
 
 void Propulsion::calcWrench()
 {
@@ -15,16 +13,16 @@ void Propulsion::calcWrench()
 
 void Propulsion::calcAdditionalData()
 {
-    // airspeed, alpha, beta
+    // airspeed
     float u_r = model->model_states.u - model->airdata.wind_x; 
     float v_r = model->model_states.v - model->airdata.wind_y;
     float w_r = model->model_states.w - model->airdata.wind_z;
     airspeed = sqrt(pow(u_r, 2) + pow(v_r, 2) + pow(w_r, 2));
 }
 
-
 BeardEngine::BeardEngine(Model *parent) : Propulsion(parent)
 {
+    initParam();
     printf("beard engine constructor\n");
 }
 
@@ -32,14 +30,12 @@ void BeardEngine::calcThrust()
 {
     float delta_t = model->control_signals.delta_t;	
     prop_wrenches.thrust = 1.0 / 2.0 * rho * s_prop * c_prop * (pow(delta_t * k_motor, 2) - pow(airspeed, 2));
-
 }
 
 void BeardEngine::calcTorque()
 {
     float delta_t = model->control_signals.delta_t;
     prop_wrenches.torque = -k_t_p * pow((k_omega * delta_t), 2);
-
 }
 
 NoEngine::NoEngine(Model *parent) : Propulsion(parent)
@@ -50,16 +46,14 @@ NoEngine::NoEngine(Model *parent) : Propulsion(parent)
 void NoEngine::calcThrust()
 {
     prop_wrenches.thrust = 0;
-
 }
 
 void NoEngine::calcTorque()
 {
     prop_wrenches.torque = 0;
-
 }
 
-void Propulsion::initParam()
+void BeardEngine::initParam()
 {
     int id = 1;   // need fix
     char paramMsg[50];
